database files refactor
